+++ packages/confirmation-modal/src/ConfirmationModal.story.tsx
@@ -66,6 +66,6 @@ function Delete() {
   );
 }
 
-storiesOf('ConfirmationModal', module)
+storiesOf('Packages/ConfirmationModal', module)
   .add('Confirm', () => <Confirm />)
   .add('Delete', () => <Delete />);