+++ packages/select/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @leafygreen-ui/select
 
+## 4.0.2
+
+### Patch Changes
+
+- Updated dependencies [f3aad7e2]
+- Updated dependencies [233ac580]
+- Updated dependencies [ba4aab15]
+- Updated dependencies [ba4aab15]
+- Updated dependencies [2cf1bc4a]
+- Updated dependencies [679b6239]
+- Updated dependencies [ef84b5fd]
+- Updated dependencies [f3aad7e2]
+- Updated dependencies [c1f9c4d4]
+  - @leafygreen-ui/button@13.0.1
+  - @leafygreen-ui/tokens@1.2.0
+  - @leafygreen-ui/lib@9.2.1
+  - @leafygreen-ui/popover@7.2.3
+
 ## 4.0.1
 
 ### Patch Changes